Rename interfaces and classes related to product salability
- Move GetProductSalableQtyInterface under InventorySalesApi module
- Move GetProductSalableQty under InventorySales module
- Move IsProductSalableInterface under InventorySalesApi module
- Move IsProductSalable under InventorySales module
- Adjust all reference to the above elements accordingly
