+++ app/code/Magento/InventorySales/Test/Integration/Stock/GetProductSalableQtyTest.php
@@ -5,12 +5,12 @@
  */
 declare(strict_types=1);
 
-namespace Magento\Inventory\Test\Integration\Stock;
+namespace Magento\InventorySales\Test\Integration\Stock;
 
 use Magento\Inventory\Model\CleanupReservationsInterface;
-use Magento\InventoryApi\Api\GetProductSalableQtyInterface;
 use Magento\InventoryApi\Api\ReservationBuilderInterface;
 use Magento\InventoryApi\Api\AppendReservationsInterface;
+use Magento\InventorySalesApi\Api\GetProductSalableQtyInterface;
 use Magento\TestFramework\Helper\Bootstrap;
 use PHPUnit\Framework\TestCase;
